Implemented cristiana and fabios suggestions
