+++ inductiva/scenarios.py
@@ -8,6 +8,7 @@ from inductiva.types import Path
 from inductiva.simulation import Simulator
 from inductiva.utils.files import resolve_path, get_timestamped_path
 from inductiva.utils.misc import split_camel_case
+from inductiva.utils.task import get_task_info, fetch_task_output
 import json
 
 
@@ -33,6 +34,15 @@ class Scenario(ABC):
         if config_filename:
             args += (config_filename,)
         return args
+
+    def download_outputs(self,
+                         output_dir: str = None):
+        """Download the outputs of an async simulation to output_dir."""
+
+        if not get_task_info(self.task_id)["status"]:
+            raise ValueError("Simulation not finished.")
+
+        fetch_task_output(self.task_id, output_dir=output_dir, return_type=None)
 
     @abstractmethod
     def gen_aux_files(self, simulator: Simulator, input_dir: Path):
